124 - Profile sections should be links
