+++ src/pages/Profile/Layout/Styles.ts
@@ -1,0 +1,13 @@
+import styled from 'styled-components';
+
+export const Container = styled.div`
+  display: grid;
+  gap: 32px;
+  grid-template-columns: 260px 1fr;
+  padding: 24px 32px;
+`;
+
+export const OutletContainer = styled.div`
+  height: 100%;
+  overflow-y: auto;
+`;
