+++ src/pages/Profile/Layout/UserDetails/Styles.ts
@@ -1,0 +1,28 @@
+import styled from 'styled-components';
+
+import UButton from 'components/Button';
+
+export const Button = styled(UButton)`
+  margin-top: 16px;
+  width: 100%;
+`;
+
+export const Container = styled.div``;
+
+export const Img = styled.img`
+  border-radius: 50%;
+  height: 100%;
+  object-fit: cover;
+  position: absolute;
+  width: 100%;
+`;
+
+export const ImgWrapper = styled.div`
+  padding-bottom: 100%;
+  position: relative;
+  width: 240px;
+`;
+
+export const Username = styled.h1`
+  margin-top: 16px;
+`;
